// urgoo_web/src/models/user/UserProfileResponseDto.ts
export interface UserProfileResponseDto {
  msg: string;
  status: number;
  firstName: string;
  lastName: string;
  balance: number;
  urgxBalance: number;
  chainAddress: string;
  regNo: string;
  phoneNo: string;
  referralCode: string;
  statusCode: number;
  statusText: string;
  idFrontConfirmedStatus: number;
  idFrontConfirmedStatusText: string;
  idBackConfirmedStatus: number;
  idBackConfirmedStatusText: string;
  selfieWithIdConfirmedStatus: number;
  selfieWithIdConfirmedStatusText: string;
  idFront: string;
  idBack: string;
  selfieWithIid: string;
  email: string;
  emailConfirmedFlag: boolean;
  pwdChangeRequiredFlag: boolean;
  nickname: string;
  fb: string;
  twitter: string;
  instagram: string;
  verifiedFlag: boolean;
  likes: number;
  seen: number;
  created: number;
  shared: number;
  profilePictureMd: string;
  profilePictureSm: string;
  coverPhoto: string;
  bio: string;
  configuredUrgooLoyaltyFlag: boolean;
}